Exercicios 1 e 2
