+++ modulo4/relacoes-sql/src/data/baseDataBase.ts
@@ -1,0 +1,14 @@
+import dotenv from "dotenv"
+import knex from "knex"
+
+dotenv.config()
+export const connection = knex({
+	client: "mysql",
+	connection: {
+    host: process.env.DB_HOST,
+    port: 3306,
+    user: process.env.DB_USER,
+    password: process.env.DB_PASS,
+    database: process.env.DB_NAME
+  }
+})
